Fix some unexpected cases of image maps.
Fix for buganizer issue 1800613: Image map not accessible.  We were attempting to cache associtations of <area> elements to their RenderImages so we would not have to search for them.  However, we were storing them in CacheBuilder.  When we search for them, we search in the <area> element's frame, even though when we built the cache, we stored it in the main frame.  Another issue arose on a page where <area> elements show up before their RenderImages.  In this case, we were storing an empty rectangle for the bounds, and caching the associations afterwards, when it was too late.  In order to clean up/simplify things, while fixing both bugs, I have changed validNode into a static function which takes the starting Frame as a parameter.  Any time we need to find the rectangle of an <area> element, we have to traverse the DOM to find its associated RenderImage.  This may be slower, but it will always get the correct answer.  In the future, we may need to investigate a universal location for caching the associations for better performance.