(1 of 2) add call to determine if picture is ready
This is a half-checkin. The other half is in frameworks/base.

Add WebViewCore::pictureReady(). This uses the last reported
progress when the picture was recorded, and checks to see if
the picture itself is blank. The caller can use this to show
a cached picture or switch to a new one.

http://b/issue?id=1802703