Shave ~60k off of libwebcore.so and make webcore_test always build.
Build libwebcore as a static library first. Then build it as a shared library
using the jni entry point as the single source file. This strips uneeded code
out of the final library.

Have webcore_test link against the static library so that the symbols are all
visible. Update webcore_test to link against the new webkit drop as well.

Rearrange the main makefile to create WEBKIT_* variables that a sub-makefile
like webcore_test can use to define includes, cflags, and libraries. Manually
add a dependency on the generated webkit source files.

Verified that svg and v8 work as well as the simulator.