use trackball 'mouse' position instead of current focus on click
Furthering investigation to separate browser focus movement from
trackball mouse position movement, use the simulated mouse
position when computing the node underneath used for mouse clicks.

The computed node is only used for area maps and lists -- usually
the (x,y) position is passed through to webkit. Since this is
re-finding the node by location and not using the node found by
navigation, it may introduce some regression.

An additional regression/investigation would be to pass the
(x,y) through regardless of whether a node is hit so that the click
is always generated.

http://b/issue?id=1866439