Update the language from "focus".
Previously, many methods, variables, etc referred to the focus.
Now, they have been changed to refer to the cursor when
appropriate.  Some references were changed to the focusCandidate,
since they may apply to either the focus or the cursor, which
is a candidate for focus.  Also removed unneeded method
unblockFocus.  Will not work without a corresponding
change in frameworks/base.