Remove obsolete notion of invalid node.
In the old navigation model, we set focus to a node, and checked to see
if that node was still valid (could have changed since we built the
navigation cache).  Then we sent a message back to the UI thread to
tell it to draw the ring differently.  In the new model, we are essentially
moving a mouse, so we do not concern ourselves with whether the node is
valid.  Here I have removed function and jni to call sendMarkNodeInvalid,
the notion of an invalid node in WebView.cpp, and the INVALID flavor
of CursorRing.  Needs a change in frameworks/base to work properly.