rebuild webview nav cache if cursor node changes
Everytime the picture set is rebuilt, the nav cache
is also built if the dom changes. This doesn't catch
javascript that changes locations of objects without
updating the dom. To handle these cases, do a hit-test
on the cursor location and see if it finds the same
node used to build the cursor cache.

This helps track the gmail message floating tool bar.