rebuild WebView nav cache on trackball move, other bugs
When the WebViewCore picture set is recomputed, check
to see if the trackball has moved since the last time.
If so, rebuild the nav cache even if dom version checking
is disabled.

If there's no focus node ready to receive a key event,
check to see if the cursor is hovering over a plugin, and
give it the event instead.

Update all cursor state in both places it could change.