Remove ANDROID_USER_GESTURE as the bug it fixed no longer applies.
I tested the original bug and found that without this change, the link works
perfectly fine. The only thing that does not work with this change is redirects
within iframes that are to non-http protocols. This is ok since iframes can
always change the parent window's location if this is the desired behavior.