Fix <select with multiple and/or size > 1.
With the latest WebKit drop we lost a hack we had to make ListBoxes
behave like MenuLists.  Rather than reviving it, now in WebViewCore
check for isListBox().  This makes a click once again open the list
so the user can choose the selection.  Also disable WebKit's drawing
of the text shown, and simply draw it ourselves, so we can draw
nothing or the first item selected.