Fix a bug where sometimes the cursor was lost.
Create a new function that sets the cursor data to the WebViewCore,
and call it each time we change the cursor.  Prior to this change,
by touching on a non-cursored, non-focused textfield, you could
set the cursor on the current CachedRoot without setting the data
on the WebViewCore.  Then, when we get a new frame cache, we
check the data on the WebViewCore (which is out of date).  Since
it does not think there was a cursor (or it thinks it was somewhere
else), we do not restore the cursor on the new cache.  This change
fixes that by ensuring that the cursor information is up to date.