Fix a WebView longpress bug.
Longpressing on a WebView while the WebTextView is up calls
WebTextView::performLongPress().  This used to be fine, but now
we may still leave the WebTextView up while the user is longpressing
on something else.  Check to ensure that the focus is the same
as the cursor before calling WebTextView's version.  Also,
send a click if the user longpresses on a different WebTextView,
so it will now behave appropriately.  Requires a change in
frameworks/base.

Provide a setter on CachedHistory to set the mousebounds, and
set it in selectBestAt, so that getSimulatedMousePosition
uses the correct information.  Also edit some comments to refer
to the cursor rather than the focus.  In WebView.cpp, use
the delete TypingCommand, making deleting and cutting work
once again.