Added createView and attachView for playing embedded videos.
Use a larger naturalSize (same as RenderReplaced intrinsic size). The real code
should report the size from the video once decoded.

Still need to implement setVisible to not show the VideoView when it isn't
actually visible. That would probably be a better spot to attach it to the view
system.