In WebTextView, check the focus before sending a key.
This fixes a bug where the DOM has moved the focus, but the WebTextView
is still sending keys to the old focus.  Requires a change in
frameworks/base.  This part of the change just gives WebView a way
to get a pointer to the currently focused node (regardless of the
cursor).