hide the cursor when the link is followed
The cursor position is preserved, so additional trackball
navigation will start from where the cursor ring was drawn
last. Hiding the cursor gets rid of confusion where the
cursor ring stays visible when clicking on a link that
does not go to a new page, or when the new page has a
link at the same physical location as the old page.