Do not merge
Pull down http://trac.webkit.org/changeset/42163 to fix the icon database crash.

I believe the issue was that m_syncThreadRunning could have been initialized to
false if the moons aligned. When the Browser quits, it closes the icon db. If
m_syncThreadRunning happened to be false, the close method would not wait for
the sync thread to complete before cleaning up (and closing) the sqlite
database. I think that caused the underlying db to be null within sqlite.