Fix a memory leak caused by the binding code.
Old code reinitialize a context when a frame loading a new page, even when the page does not have JavaScript. It also initializes an empty context when before an iframe is destroyed.

A fix is made to Chromium project (not it is in WebKit), our next merge should pick up this fix.