Add Plugin APIs for creating and manipulating a Surface.
SurfaceCallback: An interface used to receive the surface change events. The
PluginSurface then sends those events to the actual plugin.

PluginSurface: A wrapper around the native Surface object obtained from the Java
SurfaceView. This object can lock and unlock the surface and takes care of
creating, positioning and destroying the Java SurfaceView.

Changed PluginWidgetAndroid to not use SkFlipPixelRef when the drawing model is
Surface.