Initialize context if a frame had an empty global object.
When a frame navigates to a new page, the global object is detached from its
old context. However, if the new page has no JavaScript, its new context is not initialized. If another page kept a reference to the global object of the old page and tries to access the new page, it will fail because the context of new page is not initialized so the global object is connected to a real context.