Check for xml and text mime types through DOMImplementation.
This change in conjunction with a change in MimeTypeMap allows text extensions
to map to text/* mime types. Previously, a text mime type would be downloaded
instead of displayed if MimeTypeMap mapped a text extension to an unknown text
mime type. Use DOMImplementation to check for xml and text mime types since it
will compare the mime type to "text/"