Clear the WebTextView when clicking on another node.
We already cleared it in the above case, when the
user clicks on empty space.  If the user clicks
on a node, and it is not a text input, clear it.

This fixes a bug on the Google home page, where
the user has typed into the textfield, but then
clicks on "more" and is then unable to click on
Gmail, which is covered by the (invisible)
WebTextView.