Do not merge.
This will attach and detach the native thread to and from the java vm. I believe
this was the cause of a crash in donut. Sometimes, the IconDatabase needs to
notify the WebCore thread of a new Icon. These callbacks post messages through
the vm which causes AttachCurrentThread to be called but no matching
DetachCurrentThread when the thread exits.

Bug:1977972