Fix a bug in find.
Fix for http://b/issue?id=2021424.  When trying to draw the matches,
if we find that the page needs to be scrolled to put the match on
screen, we request a scroll, do a viewInvalidate and stop, so that
the matches will be drawn after the invalidate.  Unfortunately, for
some values, we do not end up scrolling, because contentToView turns
the scroll into a no op.  So we forever request scrolls, and never
draw the matches.  Fix this by getting a return value from scrollBy,
and if it returned false, go ahead and draw the matches.