Be more restrictive when expanding iframes.
If an iframe has no scrollbars and a fixed dimension, it serves no purpose to
try and expand the contents. Some sites like to use hidden iframes for
asynchronous loading and showing or expanding those iframes causes layout
problems.

Change the expansion logic slightly to only expand iframes and not contract
them. Also update calcWidth and calcHeight to check for scrollbars or a
non-fixed dimension.

BUG=2039520,2004093