Correctly applies Geolocation timeout parameter.
Currently, the timeout is started when permissions are granted and stopped
when the success callback is made, thus rendering the timeout useless. The
correct behavior is to start the timeout when the request is started, and
to stop it as soon as a position fix is obtained.

Also, the timeout should always be applied, even if no error callback is present
and the request should be stopped on timeout.

This will be upstreamed to WebKit in bug 27256.