Fix errors when compiled under g++ 4.4.1 on Ubuntu 9.10a4
This environment ends up using the C++ version of strrchr rather than the
C version. Apparently older environments use the C version of strrchr.
I'm guessing the change is due to one of the standard C++ header files
changing so that includes the new C++ version of strrchr.

The difference is that the C version of strrchr is:

char* strrchr(const char*, char);

while the C++ version of strrchr is overloaded:


const char* strrchr(const char*, char);
char* strrchr(char*, char);

This code depended on the C version to convert a const char* pointer into
a non-const char* pointer.

The fix is to change the code to work with a const char* result, and if
that is not possible, manually insert a hard cast of the return value.