fix address detection when the city and state are both valid states
This fixes detecting "4 E. 86th St New York, NY"

The old logic found 'New York' and assumed that St was the city
name. Since 86th is not a valid street suffix, the test failed.

The new logic looks for a valid street suffix instead of skipping
'city'. If it finds one, and the subsequent street suffix test
fails, it resets to search again for a valid state name while
retaining the starting point for the address (the street number).

Fixing this exposed a bug in the zip code detection code
where it dropped the first character of the actual state name.