save and restore the transparent fill when drawing replaced images
If an image node is incomplete, RenderImage::paintReplaced is triggered.
A thin frame with a transparent center sets but does not restore the
fill to transparent. Our graphics state draws transparent bitmaps until
the state gets set back to opaque.

I haven't figured out why Safari doesn't demonstrate this same bug, but
in some webkit code, the state is saved and restored around making the
fill transparent. Adding that save()/restore() pair fixes our bug as
well. But, it may be that our platform should be setting the fill
state before drawing the bitmap and is failing to do so.

This fixes http://b/issue?id=2052757