fix parent index of focus when building nav cache
When a focused node is found while walking the DOM, the
CachedFrame tree doesn't have its parent pointer set.
Fix the focus index in the parent when fixing the parent pointer.

Also fix a debug statement so that it doesn't depend on
the focus parent.

Fixes http://b/issue?id=2048186