don't short-circuit if the text selection has changed
The webkit picture is rebuilt pretty often -- for performance,
the nav cache is built less often, only when the dom version changes.

Additionally, it's rebuilt when the focus changes or moves. Add
a condition so that it is also rebuilt if the focus selection changes.

fixes http://b/issue?id=2096746