fix cracks when we draw the page in portions (i.e. a picture set)
When we record/draw the DOM N times (for speed), webkit sees what we're doing in certain places,
and sends us different rectangle coordinates. These were all being antialiased (by default).
However, when we are also zoomed, the rects now fall on fractional coordinates, and with aa,
we will double-draw the edges where those rects should have seamed.

The fix is to disable antialiasing for a class of rects that we record from webkit. We are probably
disabling for more cases than is necessary for the current bug, but knowing which ones are
"required" is tricky, and there (as yet) seems to be no down-side, since we never draw the page
rotated at a funny angle (where the rect edge would look jaggie).

http://b/issue?id=2132971&cookieId=2009268114917835
http://b/issue?id=2127763&cookieId=2009268114931860