do not merge
https://android-git.corp.google.com/g/#change,22779 chose
poorly in its checkin comment and was inadvertantly left
out of the automerge process. Here it is, along with the
original comment:

when finding left edge of text block, don't merge in tall bitmap

Zooming in to yahoo.com fails sometimes if the block of text has
a large bitmap on the left. This change prevents the bitmap from
being seen as a part of the text paragraph.

Also, cleaned up the debugging statements.

Fixes http://b/issue?id=2066236