Remove superfluous files from WebCore/svg/graphics.
This change removes the contents of the following directories, which were removed in http://trac.webkit.org/changeset/44332
- WebCore/svg/graphics/cairo
- WebCore/svg/graphics/cg
- WebCore/svg/graphics/filters/cg
- WebCore/svg/graphics/mac
- WebCore/svg/graphics/qt
- WebCore/svg/graphics/skia

These files contained Android specific edits, but these were trivial (ENABLE(FILTERS) changed to ENABLE(SVG_FILTERS)).
Due to these changes, the files were not deleted in subsequent WebKit merges.
Only WebCore/svg/graphics/skia/SVGResourceFilterSkia.cpp was used in Android Makefiles, and this was only in the V8 Makefile.

When SVG is re-enabled, the Android code will have to be updated to work with the refactoring that has taken place in WebKit, so I'm deleting these files now to ease WebKit merging and upstreaming.

Change-Id: I4cc74167193175d05bf6b0382b3059b5230cc94d