fix multiple text areas, and text areas without focus rings
Separate nodes that are hidden from nodes that disable
drawing the focus ring. If a node draws its own focus state,
other than not drawing the ring, it should be indistinguishable
from other focusable nodes.

Fix the nav cache dump to be in line with the latest webkit.

fixes http://b/issue?id=2201049