tweak select text
webkit changed the behavior of hitTestResultAtPoint; it used
to return the innermost node hit, but does so no longer.
Work around this by searching the returned node for a text
node.

Start the text detection one pixel lower. This fixes picking up
the first line of a link on m.msn.com.

Simplify the selection drawing. Share code between the arrow and
the i-beam. Center the i-beam on the arrow. Use relative path
construction to simplify the i-beam.

Add debugging.

Requires a companion change in frameworks/base

partially fixes http://b/issue?id=1667690