do not merge -- Fix a clicking bug.
original at https://android-git.corp.google.com/g/#change,31300

Remove some code that simulates a mouse click at the beginning of
a textarea and the end of a textfield.  The original goal was to
make the click change the selection to be at the beginning or end
of the field, respectively.  However, we actually make another call
which prevents this click from the selection.  Further, the
selection actually gets changed elsewhere.

Fixes http://b/issue?id=2219233