Change window focus only if the user initiated the event.
This does not work in the javascript case because the user gesture is stored in
the caller's script environment and that knowledge is not obtainable in the
receiving frame. However, this change allows targeted links to change the focus
to the correct window but prevents sites like slate.com from changing back to
the opening window.

This change will be followed by a simple revert of the previous change in the
Browser app.

Bug: 2161671