Moves get_thread_msec from JavaScriptCore/wtf/CurrentTime to WebKit/android/Timecounter.
This function is an Android addition to JavaScriptCore::WTF, but is not used anywhere in WebCore or JavaScriptCore.
Futhermore, JavaScriptCore/wtf/CurrentTime does not contain any other platform additions.
It's used primarily in WebKit/android/TimeCounter, so I've moved it there.
I've also updated the name to getThreadMsec to match WebKit style.

Change-Id: Ia4f91efd3caaafd1fb079afe6c493aefea402582