Update the Android-specific fix in StringHash::equal to match the fix recently made in webkit.org.
StringHash::equal includes an Android-specific fix for platforms which don't
provide 4-byte alignment ...

https://android-git.corp.google.com/w/?p=platform/external/webkit.git;a=commitdiff;h=05d23612a67d710479fb2359e98d7ac49a982025

This was recently fixed in webkit.org ...

http://trac.webkit.org/changeset/51006

This patch updates Android to use the fix from webkit.org to avoid future merge conflicts.

Background ...

The fix in webkit.org uses a loop, whereas the Android fix uses memcmp.
Tests by WebKit have apparently shown a loop to be faster than memcmp for small
strings. This is consistent with an equivalent fix in AtomicString ...

http://trac.webkit.org/changeset/19930

Change-Id: Ic9c0f729117e31bfdcb1bc1c28e85c048a464173