navigate preferably between children of the same parent
The nav cache attempts to take advantage of the order that
the dom is walked to know when multiple nodes have the same
parent. The old implementation doesn't always work, and a
simpler non-cached version makes more sense.

The algorithm now walks nodes until the parent has more than
one child, and assigns that parent as the 'parent group'. On
the other end, nodes with no parent group are never allowed
to get preferential matching treatment.