nav to container if it wants key events
When a text field or text area (or plugin) has another navable
element completely overlaying it, allow the trackball to move
from that inner element to the outer one.

Also, allow this to get a 'second chance' -- treat the case
when one element overlays another as a spacial nav and allow
it to retry in nav find algorithms.

fixes http://b/2266756