Provide <input> type information to Java side.
Help to fix http://b/issue?id=1890360 and http://b/issue?id=2150538

CacheBuilder.cpp:
Explicitly set isTextField to false for textareas.

CachedRoot:
Remove the code which checks to see if the textfield is a search,
since if it is, we can avoid this path altogether.

WebView:
Return a single integer which tells what type the current text
input field is.

Requires a change to frameworks/base.