Fix a rare crash when dispatching icon notifications.
I believe this happened when a WebIconDatabaseClient registered twice for
notifications. When it unregistered, it only removed the first entry. This could
lead to a crash since the second entry would be pointing to possibly deleted
memory.

Remove the mutex protection since all operations on mClients happen in the same
thread.

Add a missing include file in WebHistory found while debugging this crash.

Bug: 2279055