Updates Document::setExtraLayoutDelay to reflect recent submission to webkit.org.
setExtraLayoutDelay was an Android addition and was upstreamed to webkit.org in
https://bugs.webkit.org/show_bug.cgi?id=32875

Removes Document::extraLayoutDelay by incorporating the extra delay in Document::minimumLayoutDelay.
Removes ANDROID_MOBILE guards, as this extra code is not inhernetly specific to Android, and is a
no-op if Document::setExtraLayoutDelay is not called. This removes the last uses of ANDROID_MOBILE.

Change-Id: I1e18a994f32b29558dcf9a5c76aefab62e9fc13d