extract selected text from the picture
The deleted code extracted the selected text from the WebKit DOM
by looking for text in the render tree that matched the selection
region. This has never worked well.

The replacement approach matches the glyphs in the picture instead.

- WebViewCore.*
Remove getSelection() and supporting routines. Part of the deleted
code looked for punctuation between segments to add space characters
that might not be present in the markup. There's no equivalent in
the replacement code; it uses spacial gaps to detect the need to
add extra spaces.

- SelectText.*
Match the text to the selected region and convert the glyph to
unicode. The tricky part is that spaces need to be inserted into
the string when there are gaps in the text.

- WebView.cpp
Return the selection as a string to the java caller instead
of a region.

To convert the glyphs into text, a companion change was made to
external/skia. To update the UI thread/webkit thread messaging, a
companion change was also made to frameworks/base.

Fixes http://b/2166748