DO NOT MERGE.
If the old screen width is 0, don't bother find the
node to realign the page.

This shaves off 1/2 of the increased time in the
page_cycler due to the new pinch changes.

e.g. with moz suite.
It was 201422 ms
After pinch changes, it is 218531 ms
After this change, it is 207367 ms

Fix http://b/issue?id=2371694