Moves V8's JavaField from V8's jni_runtime to JavaClassV8 and fixes style
JavaField is script-engine specific. The JSC version of JavaField was moved out
of the script-independent file JNIBridge (previosuly named jni_runtime) in
http://trac.webkit.org/changeset/53849

This change makes the equivalent move for V8.

This will make V8's version of jni_runtime script-independent, allowing it to be
merged with JNIBridge in WebCore in the future.

This change is being upstreamed to webkit.org in
https://bugs.webkit.org/show_bug.cgi?id=34166

Change-Id: I7b35bfe4e8e7820623a824028eb64d06421bdb6b