Moves the Geolocation position cache out of the Geolocation object.
The position cache is an implementation detail of the Geolocation object,
but is being moved to its own file to keep files to a reasonable size.
Also updates some names and fixes soem style.

This will be upstreamed in
https://bugs.webkit.org/show_bug.cgi?id=34084

The maximumAge code in the Geolocation object,
which may make use of this caching, is being upstreamed in
https://bugs.webkit.org/show_bug.cgi?id=30676

Change-Id: I083556de732bd7be240a75da5369117aabc50e33