Merge webkit.org at r54127 : Fix conflict with addition of virtual destructor to Visitor class
Android had added a virtual destructor to DOMWrapperMap::Vistor
See https://android-git.corp.google.com/g/#change,21263

DOMWrapperMap::Vistor was removed and WeakReferenceMap::Visitor was added in
http://trac.webkit.org/changeset/51998
DOMWrapperMap inherits from WeakReferenceMap.

The virtual desctructor was upstreamed to webkit.org for WeakReferenceMap::Visitor in
http://trac.webkit.org/changeset/53401
Note that the message in the ChangeLog is incorrect.

Change-Id: Ia801155d0b42100e91b891b0568486fc76d037e1